+++ CHANGELOG.md
@@ -5,11 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.6] - 2023-02-20
+- add parameter position to **resetPosition()**
+- add reset of \_lastPosition to really nullify the position.
+- update readme.md
+
+
 ## [0.3.5] - 2023-02-01
 - update GitHub actions
 - update license 2023
 - update readme.md
-
 
 ## [0.3.4] - 2022-12-22
 - fix #26 edges problem of the experimental cumulative position (CP).